initial separation of container and template
